Move new tests to data_plane.go
