+++ test/rekt/features/apiserversource/webhook_validation_smoke.go
@@ -19,21 +19,13 @@ package apiserversource
 import (
 	"context"
 
-	"github.com/cloudevents/sdk-go/v2/test"
 	"github.com/stretchr/testify/assert"
-
-	rbacv1 "k8s.io/api/rbac/v1"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	v1 "knative.dev/eventing/pkg/apis/sources/v1"
-	"knative.dev/eventing/test/rekt/resources/account_role"
 	"knative.dev/eventing/test/rekt/resources/apiserversource"
 	duckv1 "knative.dev/pkg/apis/duck/v1"
-	"knative.dev/reconciler-test/pkg/eventshub"
-	eventasssert "knative.dev/reconciler-test/pkg/eventshub/assert"
 	"knative.dev/reconciler-test/pkg/feature"
-	"knative.dev/reconciler-test/pkg/manifest"
-	"knative.dev/reconciler-test/resources/svc"
 )
 
 func CreateWithInvalidSpec() *feature.Feature {
@@ -43,44 +35,6 @@ func CreateWithInvalidSpec() *feature.Feature {
 
 	f.Stable("ApiServerSource webhook").
 		Must("reject invalid spec on resource creation", createApiServerSourceWithInvalidSpec(srcname))
-
-	return f
-}
-
-func SendsEventsWithSinkRef() *feature.Feature {
-	source := feature.MakeRandomK8sName("apiserversource")
-	sink := feature.MakeRandomK8sName("sink")
-	f := feature.NewFeature()
-
-	f.Setup("install sink", eventshub.Install(sink, eventshub.StartReceiver))
-
-	sacmName := feature.MakeRandomK8sName("apiserversource")
-	f.Setup("Create Service Account for ApiServerSource",
-		account_role.Install(sacmName,
-			account_role.WithRole(sacmName+"-clusterrole"),
-			account_role.WithRules(rbacv1.PolicyRule{
-				APIGroups: []string{""},
-				Resources: []string{"events"},
-				Verbs:     []string{"get", "list", "watch"},
-			}),
-		))
-
-	cfg := []manifest.CfgFn{
-		apiserversource.WithServiceAccountName(sacmName),
-		apiserversource.WithEventMode(v1.ResourceMode),
-		apiserversource.WithSink(svc.AsKReference(sink), ""),
-		apiserversource.WithResources(v1.APIVersionKindSelector{
-			APIVersion: "v1",
-			Kind:       "Event",
-		}),
-	}
-
-	f.Setup("install ApiServerSource", apiserversource.Install(source, cfg...))
-	f.Requirement("ApiServerSource goes ready", apiserversource.IsReady(source))
-
-	f.Stable("ApiServerSource as event source").
-		Must("delivers events",
-			eventasssert.OnStore(sink).MatchEvent(test.HasType("dev.knative.apiserver.resource.add")).AtLeast(1))
 
 	return f
 }
